+++ excalidraw-app/collab/CollabError/CollabError.tsx
@@ -48,4 +48,3 @@ const CollabError = ({ collabError }: { collabError: ErrorIndicator }) => {
 CollabError.displayName = "CollabError";
 
 export default CollabError;
-