+++ excalidraw-app/components/Workspace/UserProfileDialog/index.ts
@@ -1,2 +1,1 @@
 export { UserProfileDialog, default } from "./UserProfileDialog";
-