+++ excalidraw-app/components/Workspace/WorkspaceMainContent/index.ts
@@ -1,3 +1,2 @@
 export { WorkspaceMainContent, default } from "./WorkspaceMainContent";
 export type { WorkspaceMainContentProps } from "./WorkspaceMainContent";
-