+++ excalidraw-app/components/Workspace/WorkspaceSidebarTrigger/index.ts
@@ -1,2 +1,1 @@
 export { WorkspaceSidebarTrigger, default } from "./WorkspaceSidebarTrigger";
-