+++ excalidraw-app/share/ShareDialog/ShareDialog.tsx
@@ -401,4 +401,3 @@ export const ShareDialog = (props: {
     />
   );
 };
-